+++ CHANGELOG.md
@@ -30,6 +30,10 @@ Line wrap the file at 100 chars.                                              Th
 - The app will have it's window resized correctly when display scaling settings are changed. This
  should also fix bad window behaviour on startup.
 
+### Fixed
+#### Linux
+- Fixed systemd-resolved DNS management.
+
 
 ## [2018.4-beta2] - 2018-10-08
 ### Added